Removed field "type" from a distribution update.
