+++ hawkbit-repository/hawkbit-repository-api/src/main/java/org/eclipse/hawkbit/repository/builder/DistributionSetUpdate.java
@@ -8,13 +8,10 @@
  */
 package org.eclipse.hawkbit.repository.builder;
 
-import java.util.Optional;
-
 import javax.validation.constraints.NotNull;
 import javax.validation.constraints.Size;
 
 import org.eclipse.hawkbit.repository.model.DistributionSet;
-import org.eclipse.hawkbit.repository.model.DistributionSetType;
 import org.eclipse.hawkbit.repository.model.NamedEntity;
 import org.eclipse.hawkbit.repository.model.NamedVersionedEntity;
 
@@ -44,22 +41,6 @@ public interface DistributionSetUpdate {
      * @return updated builder instance
      */
     DistributionSetUpdate description(@Size(max = NamedEntity.DESCRIPTION_MAX_SIZE) String description);
-
-    /**
-     * @param typeKey
-     *            for {@link DistributionSet#getType()}
-     * @return updated builder instance
-     */
-    DistributionSetUpdate type(@Size(min = 1, max = DistributionSetType.KEY_MAX_SIZE) @NotNull String typeKey);
-
-    /**
-     * @param type
-     *            for {@link DistributionSet#getType()}
-     * @return updated builder instance
-     */
-    default DistributionSetUpdate type(final DistributionSetType type) {
-        return type(Optional.ofNullable(type).map(DistributionSetType::getKey).orElse(null));
-    }
 
     /**
      * @param requiredMigrationStep
